# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "\"Bearer\" followed by a space and JWT token (Specify Trade vs Non Trade JWt token).",
        "title": "Go Example Project",
        "contact": {
            "name": "Add Team Name",
            "email": "Team DL or Slack Group Name"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/counter/create": {
            "post": {
                "description": "Creates a new counter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "counter"
                ],
                "summary": "Creates a new counter",
                "operationId": "createCounter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "counter key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/counter/current": {
            "get": {
                "description": "Get the current value of counter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "counter"
                ],
                "summary": "Get the current value of counter",
                "operationId": "currentCount",
                "parameters": [
                    {
                        "type": "string",
                        "description": "counter key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CounterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/counter/decrement": {
            "put": {
                "description": "Decrement an existing counter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "counter"
                ],
                "summary": "Decrement an existing counter",
                "operationId": "decrementCounter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "counter key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/counter/increment": {
            "put": {
                "description": "Increment an existing counter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "counter"
                ],
                "summary": "Increment an existing counter",
                "operationId": "incrementCounter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "counter key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/fullName": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Gets the full name from the first name and last name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fullName"
                ],
                "summary": "Gets the full name from the first name and last name",
                "operationId": "fullName",
                "parameters": [
                    {
                        "description": "first name and last name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FullNameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FullNameResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/maths": {
            "post": {
                "description": "Returns a op b",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "asyncMaths"
                ],
                "summary": "Does a maths op in a async way",
                "operationId": "asyncMaths",
                "parameters": [
                    {
                        "description": "a op b",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MathsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MathsJobResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.MathsJobResponse"
                        }
                    }
                }
            }
        },
        "/maths/{id}": {
            "get": {
                "description": "GGet the status (and result if completed) of the job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "asyncMaths"
                ],
                "summary": "Get the status (and result if completed) of the job",
                "operationId": "getJobStatus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "counter key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jobs.JobResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/jobs.JobResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jobs.JobResult"
                        }
                    }
                }
            }
        },
        "/moxy": {
            "get": {
                "description": "Get the moxy response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moxy"
                ],
                "summary": "Get the moxy response",
                "operationId": "moxy",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MoxyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "jobs.JobResult": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "res": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.CounterResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
        "models.FullNameRequest": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string",
                    "example": "Nitesh"
                },
                "lastName": {
                    "type": "string",
                    "example": "Jain"
                }
            }
        },
        "models.FullNameResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Nitesh Jain"
                }
            }
        },
        "models.MathsJobResponse": {
            "type": "object",
            "properties": {
                "res": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.MathsRequest": {
            "type": "object",
            "properties": {
                "a": {
                    "type": "integer"
                },
                "b": {
                    "type": "integer"
                },
                "op": {
                    "type": "string"
                }
            }
        },
        "models.MoxyResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}